% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{abr.names}
\alias{abr.names}
\title{abr.names shorten long column names}
\usage{
abr.names(tab)
}
\arguments{
\item{tab}{Data frame for which columns should be renamed}
}
\value{
Data frame with shortened column names
}
\description{
abr.names shorten long column names
}
