% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{AddWeatherParameters}
\alias{AddWeatherParameters}
\title{Title}
\usage{
AddWeatherParameters(df, ShiftYears = FALSE, StartMonth = 1)
}
\arguments{
\item{df}{data frame with weather data}
}
\value{
data frame with additional weather parameters
CumRain, CumRad, TempSum, PT, cPT, Rn, ra, rc, pETP, cumETP, climWbal, cumWbal
}
\description{
Title
}
\examples{
AddWeatherParameters(df)
}
