% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{air_saturation_deficit}
\alias{air_saturation_deficit}
\title{calculation of air saturation deficit of water vapour}
\usage{
air_saturation_deficit(temperature, relative_humidity)
}
\arguments{
\item{temperature}{air temperature}

\item{relative_humidity}{relative humidity}
}
\value{
air_saturation_deficit [mbar]
}
\description{
calculation of air saturation deficit of water vapour
}
