% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{clean_plotly_leg}
\alias{clean_plotly_leg}
\title{Utility function to clean up plotly legend entries}
\usage{
clean_plotly_leg(.plotly_x, .extract_str)
}
\arguments{
\item{.extract_str}{}
}
\description{
Utility function to clean up plotly legend entries
}
