% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{Copy_DWD_ZipFiles}
\alias{Copy_DWD_ZipFiles}
\title{Copy_DWD_ZipFiles makes a local copy of DWD weather data as zip files}
\usage{
Copy_DWD_ZipFiles(DWD_ftp_, LocalCopy_DWD_ftp_)
}
\arguments{
\item{DWD_ftp_}{ftp address of DWD  weather data}

\item{LocalCopy_DWD_ftp_}{directory for storing local copy of DWD data}
}
\value{
Nothing
}
\description{
Copy_DWD_ZipFiles makes a local copy of DWD weather data as zip files
}
