% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{CreateScenarioWeather}
\alias{CreateScenarioWeather}
\title{CreateScenarioWeather}
\usage{
CreateScenarioWeather(
  df = NULL,
  scenYear = 2024,
  StartMonthSim = 9,
  startYear = 2004,
  ScenarioStartDay = NULL,
  EndYear = 2024,
  site_key = "L21"
)
}
\arguments{
\item{df}{data frame with weather data}

\item{scenYear}{A year for which a scenario weather data is created}

\item{StartMonthSim}{A month for the start of the simulation scenario}

\item{startYear}{A year for the start of the simulation scenario}

\item{ScenarioStartDay}{A day for the start of the simulation scenario}

\item{EndYear}{A year for the end of the simulation scenario}

\item{site_key}{A key for the site}
}
\value{
A data frame with scenario weather data
}
\description{
CreateScenarioWeather
}
