% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{delta_f}
\alias{delta_f}
\title{Title delta_f}
\usage{
delta_f(sat_vap_press, Temp)
}
\arguments{
\item{sat_vap_press}{saturated vapour pressure [mbar]}

\item{Temp}{Temperature [°C]}
}
\value{
delta_f [mbar/K]
}
\description{
Title delta_f
}
\examples{
delta_f(20, 10)
}
