% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{dens_air}
\alias{dens_air}
\title{dens_air function for calculating air density [Kg/m3]}
\usage{
dens_air(Temp)
}
\arguments{
\item{Temp}{[°C]}
}
\value{
air density [Kg/m3]
}
\description{
dens_air function for calculating air density [Kg/m3]
}
\examples{
dens_air(20)
}
