% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{displacement_height}
\alias{displacement_height}
\title{Function displacement_height (zero plane displacement height)}
\usage{
displacement_height(crop_height)
}
\arguments{
\item{crop_height}{}
}
\value{
displacement_height [m]
}
\description{
Function displacement_height (zero plane displacement height)
}
\examples{
displacement_height(0.1)
}
