% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{DWDisMeasured}
\alias{DWDisMeasured}
\title{Title DWDisMeasured}
\usage{
DWDisMeasured(df, Maxpct_miss = 0.3)
}
\arguments{
\item{weather_historic}{historic weather data}
}
\value{
data frame with DWD station info for selected stations
}
\description{
Title DWDisMeasured
}
