% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{estSg_S0}
\alias{estSg_S0}
\title{estSg_S0: calculation of daily solar radiation from relative sunshine}
\usage{
estSg_S0(RelSun, Month)
}
\arguments{
\item{RelSun}{Relative sunshine duration of a particular day of year}

\item{Month}{Month of the year}
}
\value{
solar radiation per day [MJ/m2/d]
}
\description{
estSg_S0: calculation of daily solar radiation from relative sunshine
}
\details{
function for calculation of daily solar radiation from relative sunshine hours ####
the function uses an empirical model calibrated from DWD data "mod.Angstroem"
}
