% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{Get_ZipLists_Station_ids}
\alias{Get_ZipLists_Station_ids}
\title{Title Get_ZipLists_Station_ids}
\usage{
Get_ZipLists_Station_ids(
  dataperiod = "recent",
  loadnew = T,
  localStationdata_fn = "stationdata.RData"
)
}
\arguments{
\item{dataperiod}{"recent", "historical", "both"}

\item{loadnew}{either true or false}

\item{localStationdata_fn}{name of the local file with the station meta data}
}
\value{
list with stationlist, ziplist, histziplist, recentziplist
}
\description{
Title Get_ZipLists_Station_ids
}
