% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{getcombinedDWDRain}
\alias{getcombinedDWDRain}
\title{getcombinedDWDRain gets the combined (historic & recent) rain data for a DWD station}
\usage{
getcombinedDWDRain(
  DWDRain_content,
  station,
  local = F,
  recent_repository,
  historical_repository
)
}
\arguments{
\item{station}{The ID of the station in 5 digits but as character}

\item{local}{Option to use local stored zip-files}

\item{recent_repository}{The ftp address of the recent zip files or the local directory}

\item{historical_repository}{The ftp address of the historical zip files or the local directory}

\item{DWD_content}{A list of the DWD station}
}
\value{
data frame with DWD observations historical + recent
}
\description{
getcombinedDWDRain gets the combined (historic & recent) rain data for a DWD station
}
\examples{
getcombinedDWDRain(DWDRain_content, "00044", local=F, DWDRain_ftp_recent, DWDRain_ftp_historical)
}
