% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{getcombinedDWDWeather}
\alias{getcombinedDWDWeather}
\title{getcombinedDWDWeather gets the combined (historic & recent) weather data for a DWD station}
\usage{
getcombinedDWDWeather(
  DWD_content,
  station,
  local = F,
  recent_repository,
  historical_repository
)
}
\arguments{
\item{DWD_content}{A list structure of the DWD stations}

\item{station}{The ID of the station in 5 digits but as character}

\item{local}{Option to use local stored zip-files}

\item{recent_repository}{The ftp address of the recent zip files or the local directory}

\item{historical_repository}{The ftp address of the historical zip files or the local directory}
}
\value{
data frame with DWD observations historical + recent
}
\description{
getcombinedDWDWeather gets the combined (historic & recent) weather data for a DWD station
}
