% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{getDWDContent}
\alias{getDWDContent}
\title{getDWDContent gets the content (filenames etc.) of a DWD ftp repository}
\usage{
getDWDContent(repository, quiet = T, LoadMetaData = F)
}
\arguments{
\item{repository}{ftp address of repository with DWD weather data}

\item{quiet}{Give echo or not}
}
\value{
List with 1 data frame and one array of ZIPIDs
stationlist = data frame with the columns: "Stations_id"   "von_datum"     "bis_datum"     "Stationshoehe" "geoBreite"     "geoLaenge"
"Stationsname"  "Bundesland"
filelist: all filenames of the not zipped files repository
ziplist: names of the zip files in the repository
zipID: 5 digit ZipIDs
}
\description{
getDWDContent gets the content (filenames etc.) of a DWD ftp repository
}
