% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{getDWDRainContent}
\alias{getDWDRainContent}
\title{getDWDRainContent gets the content of the DWD ftp repository for additional rain data}
\usage{
getDWDRainContent(repository, quiet = T)
}
\arguments{
\item{repository}{ftp address of DWD repository for additional rain data}

\item{quiet}{option to echo off}
}
\value{
List with 1 data frame and one array of ZIPIDs
stationlist = data frame with the columns: "Stations_id"   "von_datum"     "bis_datum"     "Stationshoehe" "geoBreite"     "geoLaenge"
"Stationsname"  "Bundesland"
filelist: all filenames of the not zipped files repository
ziplist: names of the zip files in the repository
zipID: 5 digit ZipIDs as characer vector
}
\description{
getDWDRainContent gets the content of the DWD ftp repository for additional rain data
}
