% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{getDWDRainStationList}
\alias{getDWDRainStationList}
\title{getDWDRainStationList gets a station list from the DWD ftp repository
without ziplists and addtionally returns historical and recent stationlist (metadata)
as well as filelists, ziplists and zipIDs}
\usage{
getDWDRainStationList(historical, recent)
}
\arguments{
\item{historical}{ftp address of historical rain data}

\item{recent}{ftp address of recent rain data}
}
\value{
list of stations, recent, historical objects
stations: merged data frame of all weather stations with columns
"Stations_id"   "von_datum"     "bis_datum"     "Stationshoehe"
"geoBreite"     "geoLaenge" "Stationsname"  "Bundesland"
recent and historical: lists with
stationlist = data frame with the columns: "Stations_id"   "von_datum"     "bis_datum"     "Stationshoehe" "geoBreite"     "geoLaenge"
"Stationsname"  "Bundesland"
filelist: all filenames of the not zipped files repository
ziplist: names of the zip files in the repository
zipID: 5 digit ZipIDs
}
\description{
getDWDRainStationList gets a station list from the DWD ftp repository
without ziplists and addtionally returns historical and recent stationlist (metadata)
as well as filelists, ziplists and zipIDs
}
\examples{
getDWDRainStationList("ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/daily/kl/recent/",
}
