% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{GetForecastDataForStationList}
\alias{GetForecastDataForStationList}
\title{Title}
\usage{
GetForecastDataForStationList(fn)
}
\arguments{
\item{fn}{file name of the stationlist file}
}
\value{
a data frame with the forecast data for all station in the list
}
\description{
Title
}
