% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{getHUMEWeather}
\alias{getHUMEWeather}
\title{getHUMEWeather converts the DWD weather data into the HUME format}
\usage{
getHUMEWeather(RadWeather)
}
\arguments{
\item{RadWeather}{Data frame with augmented DWD data}
}
\value{
Data frame in HUME weather format
}
\description{
getHUMEWeather converts the DWD weather data into the HUME format
}
