% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{getRadWeather}
\alias{getRadWeather}
\title{getRadWeather calculates radiation from DWD weather data}
\usage{
getRadWeather(WeatherTab)
}
\arguments{
\item{WeatherTab}{Data frame with DWD data including sunshine duration without radiation}
}
\value{
Data frame with radiation in [MJ.m-2.d-1] and the other columns
}
\description{
getRadWeather calculates radiation from DWD weather data
}
\details{
also saturation deficit is calculated
}
