% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{GetRainData_selection}
\alias{GetRainData_selection}
\title{GetRainData_selection gets the rain data for the selected stations
by downloading the data from the DWD ftp server}
\usage{
GetRainData_selection(
  stations_selected,
  DWDRain_content,
  repository = DWDRain_ftp_historical,
  startdate = "1990-01-01"
)
}
\arguments{
\item{stations_selected}{selected stations}

\item{DWDRain_content}{list structure of available DWD additional rain data stations,
must be either DWDRain_content$recent or DWDRain_content$historical}

\item{startdate}{start date for selection of data}
}
\value{
data frame with weather data for selected stations
}
\description{
GetRainData_selection gets the rain data for the selected stations
by downloading the data from the DWD ftp server
}
\examples{
GetWeatherData_selection(stations_selected, DWD_content, local=F, startdate="1990-01-01")
}
