% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{getsingleDWDWeather}
\alias{getsingleDWDWeather}
\title{Title getsingleDWDWeather loads data for a single weather station}
\usage{
getsingleDWDWeather(station, ziplist, repository, quiet = T)
}
\arguments{
\item{station}{5 digit ID of the weather station as character}

\item{ziplist}{Character array with the ZIP-Filenames}

\item{repository}{ftp Repository address or local directory}

\item{quiet}{echo on/off}
}
\value{
data frame with weather data from DWD for either historical or recent zip files
}
\description{
Title getsingleDWDWeather loads data for a single weather station
}
\details{
This is a function which loads data for a single weather station either for
recent or historical weather data. The data are retrieved in the DWD format.
The measurement height for wind speed is taken from the meta data of the station and
added to the data frame and wind speed is corrected to a standard height of 10 m.
}
