% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{GetWeatherData_selection}
\alias{GetWeatherData_selection}
\title{GetWeatherData_selection gets the weather data for the selected stations
by downloading the data from the DWD ftp server}
\usage{
GetWeatherData_selection(
  stations_selected,
  DWD_content,
  repository = DWD_ftp_historical,
  startdate = "1990-01-01"
)
}
\arguments{
\item{stations_selected}{selected stations}

\item{DWD_content}{list structure of available DWD stations,
must be either DWD_content$recent or DWD_content$historical}

\item{startdate}{start date for selection of data}
}
\value{
data frame with weather data for selected stations
}
\description{
GetWeatherData_selection gets the weather data for the selected stations
by downloading the data from the DWD ftp server
}
