% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{GetWeatherData_selection_fst}
\alias{GetWeatherData_selection_fst}
\title{GetWeatherData_selection_fst}
\usage{
GetWeatherData_selection_fst(
  stations_selected,
  DWD_content,
  repository = DWD_ftp_recent,
  startdate = "1990-01-01"
)
}
\arguments{
\item{stations_selected}{selected stations}

\item{DWD_content}{list structure of available DWD stations}

\item{startdate}{start date for selection of data}
}
\value{
data frame with weather data for selected stations in fst format
}
\description{
GetWeatherData_selection_fst
}
\examples{
GetWeatherData_selection_fst(stations_selected, DWD_content, local=F, startdate="1990-01-01")
}
