% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{InterpolateToFST}
\alias{InterpolateToFST}
\title{Title InterpolateToFST Interpolates weather data for a list of locations}
\usage{
InterpolateToFST(
  core_weather,
  df.locations,
  DWD_content,
  DWDRain_content,
  startdate,
  recent = TRUE
)
}
\arguments{
\item{core_weather}{data frame with the core DWD data}

\item{df.locations}{data frame with the locations for which the weather data should be interpolated}

\item{DWD_content}{Meta data for the DWD weather stations}

\item{DWDRain_content}{Meta data for the DWD rain stations}

\item{startdate}{start date for the interpolation}

\item{recent}{logical, if TRUE use recent data, if FALSE use historical data}
}
\value{
a data frame with the interpolated weather data
}
\description{
Title InterpolateToFST Interpolates weather data for a list of locations
}
