% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{InterpolateWeatherData}
\alias{InterpolateWeatherData}
\title{InterpolateWeatherData interpolates the weather data from the selected station to the location}
\usage{
InterpolateWeatherData(
  station_selected,
  RainStation_selected = NULL,
  weather_dat,
  RainData = NULL,
  maxError = 2,
  startdate = "1990-01-01"
)
}
\arguments{
\item{weather_dat}{weather data from the selected station in DWD format including the distance and
height difference to the location for which to interpolate the data}

\item{maxError}{maximum deviation for a parameter from the average of the station
to be included in the interpolation}
}
\value{
interpolated data in the HUME formate
}
\description{
InterpolateWeatherData interpolates the weather data from the selected station to the location
}
\examples{
InterpolateWeatherData(station_selected, RainStation_selected, weather_dat, RainData, maxError=2)
}
