% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{LoadDWDDatafromStationlist}
\alias{LoadDWDDatafromStationlist}
\title{LoadDWDDatafromStationlist}
\usage{
LoadDWDDatafromStationlist(
  station_selected,
  RainStation_selected = NULL,
  DWD_content = NULL,
  DWDRain_content = NULL,
  weather_historic = NULL,
  weather_recent = NULL,
  local = F,
  StartYear = 1990,
  maxError = 2,
  Max_Height_diff = 100
)
}
\arguments{
\item{station_selected}{# a data frame with selected stations}

\item{RainStation_selected}{# a data frame with selected rain stations}

\item{weather_historic}{# a data frame with historical weather data}

\item{weather_recent}{# a data frame with recent weather data}

\item{StartYear}{# start year for data selection}

\item{maxError}{# maximum deviation of monthly parameter value
from average of stations as multiples of standard deviation}

\item{Max_Height_diff}{# maximum difference from location to station to which stations are included}
}
\value{
interpolated weather data in HUME format including radiation data
}
\description{
LoadDWDDatafromStationlist
}
\examples{
LoadDWDDatafromStationlist(station_selected, RainStation_selected, DWD_content, DWDRain_content, weather_historic, weather_recent, local = F, StartYear= 1990, maxError = 2, Max_Height_diff = 100)

}
