% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{makeplot}
\alias{makeplot}
\title{makeplot makes time series plots from a data frame with Time in the first column as Excel date
and the parameter to be plotted in another selected column}
\usage{
makeplot(
  df,
  parameter,
  BaseSize = 18,
  ylabel = "",
  SelYear = 0,
  ShiftYears = F,
  StartMonth = 9,
  smoothing = F,
  span = 0.5
)
}
\arguments{
\item{df}{data frame with weather/simulation data}

\item{parameter}{Column to be plotted}

\item{BaseSize}{Base char size for the plot}

\item{ylabel}{The y label string}

\item{SelYear}{The year to be highlighted}

\item{ShiftYears}{Shift the years for autumn sown crops}

\item{StartMonth}{The start month for the cultivation year}

\item{smoothing}{Smooth the historic data with a loess function}
}
\value{
A ggplot object
}
\description{
makeplot makes time series plots from a data frame with Time in the first column as Excel date
and the parameter to be plotted in another selected column
}
\examples{
makeplot(df, "TMPM", BaseSize=18, ylabel="Temperatur [°C]", SelYear=2020, ShiftYears=T, StartMonth=9, smoothing=T, span=0.5)
}
