% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{makeplot2}
\alias{makeplot2}
\title{makeplot2 a function to plot scenarios}
\usage{
makeplot2(dfhist, dfscen, parameter, BaseSize = 18, ylabel = "", SelYear = 0)
}
\arguments{
\item{dfhist}{data frame with historic data}

\item{dfscen}{data frame with scen data}

\item{parameter}{Column to be plotted}

\item{BaseSize}{Base char size for the plot}

\item{ylabel}{The y label string}

\item{SelYear}{The year to be highlighted}
}
\value{
A ggplot object
}
\description{
makeplot2 a function to plot scenarios
}
\examples{
makeplot2(dfhist, dfscen, "TMPM", BaseSize=18, ylabel="Temperature", SelYear=2020)
}
