% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{makeScenarioplot}
\alias{makeScenarioplot}
\title{makeScenarioplot makes time series plots from a data frame with Time in the first column as Excel date
and the parameter to be plotted in another selected column}
\usage{
makeScenarioplot(
  df_hist,
  df_scen,
  parameter,
  BaseSize = 18,
  ylabel = "",
  SelYear = 0,
  ShiftYears = F,
  StartMonth = 9,
  StartScenarioDate = NULL,
  smoothing = F,
  span = 0.5,
  plotly = F
)
}
\arguments{
\item{df_hist}{data frame with historic weather/simulation data}

\item{df_scen}{data frame with scenario weather/simulation data}

\item{parameter}{Column to be plotted}

\item{BaseSize}{Base char size for the plot}

\item{ylabel}{The y label string}

\item{SelYear}{The year to be highlighted}

\item{ShiftYears}{Shift the years for autumn sown crops}

\item{StartMonth}{The start month for the cultivation year}

\item{StartScenarioDate}{The start date for the scenario, i.e. the first day with unkown weather}

\item{smoothing}{Smooth the historic data with a loess function}

\item{span}{The span for the loess function}

\item{plotly}{Use plotly for interactive plots}
}
\value{
A ggplot object
}
\description{
makeScenarioplot makes time series plots from a data frame with Time in the first column as Excel date
and the parameter to be plotted in another selected column
}
