% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{Penman}
\alias{Penman}
\title{Penman function for calculating potential evapotranspiration}
\usage{
Penman(Temp, Sat_def, Net_beam, delta, gamma, l_h_v_water, ra, rc)
}
\arguments{
\item{Temp}{[°C]}

\item{Sat_def}{[mbar] or [hPa]}

\item{Net_beam}{[J/m2*s]}

\item{delta}{[mbar/K]}

\item{gamma}{[mbar/K]}

\item{l_h_v_water}{[J/Kg]}

\item{ra}{aerodynamic resistance [s/m]}

\item{rc}{bulk-canopy resistance [s/m]}
}
\value{
potential evapotranspiration [mm/d]
}
\description{
Penman function for calculating potential evapotranspiration
}
\examples{
Penman(20, 10, 100, 0.1, 0.1, 2.477 * 1E6, 10, 10)
}
