% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{R_nl}
\alias{R_nl}
\title{Long wave net radiation for daily time steps}
\usage{
R_nl(
  Tmax_k,
  Tmin_k,
  ea,
  Rs,
  Rs0,
  sigma = 4.903e-09,
  a1 = 0.34,
  a2 = 0.14,
  ac = 1.35,
  bc = 0.35
)
}
\arguments{
\item{Tmax_k}{maximum temperature [K]}

\item{Tmin_k}{minimum temperature [K]}

\item{ea}{actual vapor pressure [kPa]}

\item{Rs}{global radiation [MJ/m2/d]}

\item{Rs0}{clear sky radiation [MJ/m2/d]}

\item{sigma}{Stefan-Boltzmann constant [MJ/(m2*K4*d)]}
}
\value{
long wave backward radiation [MJ/m2/d]
}
\description{
Long wave net radiation for daily time steps
}
