% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{ra_f}
\alias{ra_f}
\title{ra_f function for calculating aerodynamic resistance}
\usage{
ra_f(wind_speed, crop_height, measure_height, f_ra_funct = "Monteith-Unsworth")
}
\arguments{
\item{wind_speed}{[m]}

\item{crop_height}{[m]}

\item{measure_height}{[m]}

\item{f_ra_funct}{[character] "Monteith-Unsworth" or 'Thom-Oliver'}
}
\value{
aerodynamic resistance [s/m]
}
\description{
ra_f function for calculating aerodynamic resistance
}
\examples{
ra_f(1, 0.1, 2, "Monteith-Unsworth")
}
