% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{rc_f}
\alias{rc_f}
\title{Title rc_f}
\usage{
rc_f(rc0, LAI)
}
\arguments{
\item{rc0}{canopy resistance under well watered condition [s/m]}

\item{LAI}{leaf area index [-]}
}
\value{
canopy resistance rc_f [s/m]
}
\description{
Title rc_f
}
\examples{
rc_f(50, 4)
}
