% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{rc_f_vectorized}
\alias{rc_f_vectorized}
\title{Title}
\usage{
rc_f_vectorized(rc0, LAI)
}
\arguments{
\item{rc0}{canopy resistance under well watered condition [s/m]}

\item{LAI}{leaf area index [-]}
}
\value{
canopy resistance rc_f_vectorized [s/m]
}
\description{
Title
}
