% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{RelRad}
\alias{RelRad}
\title{RelRad: calculation of daily solar radiation from relative sunshine
the model has been fitted with observations from the DWD (relative sunshine and global radiation)}
\usage{
RelRad(RelSun = 0.8, Month = 6)
}
\arguments{
\item{RelSun}{Relative sunshine duration of a particular day of year}

\item{Month}{Month of the year}
}
\value{
Global radiation in [MJ.m-2.d-1]
}
\description{
RelRad: calculation of daily solar radiation from relative sunshine
the model has been fitted with observations from the DWD (relative sunshine and global radiation)
}
\details{
New function for calculation of daily solar radiation from relative sunshine hours ####
the function uses directly the parameters  from  "mod.Angstroem"
}
\examples{
RelRad(0.8, 6)
}
