% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{rename.Rain}
\alias{rename.Rain}
\title{rename.Rain renames the columns of the DWD rain data}
\usage{
\method{rename}{Rain}(DWDRain)
}
\arguments{
\item{DWDRain}{data frame with original DWD rain data}
}
\value{
data frame with readable column names
}
\description{
rename.Rain renames the columns of the DWD rain data
}
\examples{
rename.Rain(DWDRain)
}
