% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{RenameDWDRain}
\alias{RenameDWDRain}
\title{RenameDWDRain renames the columns of the DWD rain data}
\usage{
RenameDWDRain(DWDRain)
}
\arguments{
\item{DWDRain}{data frame with original DWD rain data}
}
\value{
data frame with readable column names
}
\description{
RenameDWDRain renames the columns of the DWD rain data
}
