% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{RenameDWDWeather}
\alias{RenameDWDWeather}
\title{Rename columns of DWD weather data}
\usage{
RenameDWDWeather(DWDWeather)
}
\arguments{
\item{DWDWeather}{Data frame with original DWD weather data}
}
\value{
Data frame with readable column names
}
\description{
Rename columns of DWD weather data
}
