% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderImportFunctions.R
\name{render_Wetter_Analysis}
\alias{render_Wetter_Analysis}
\title{render_Wetter_Analysis}
\usage{
render_Wetter_Analysis(
  Standort,
  geoBreite,
  geoLaenge,
  Hoehe_m,
  radius,
  minstations,
  startdate,
  maxError,
  set_title,
  out_file
)
}
\arguments{
\item{Standort}{name of the site for which weather data should be estimated}

\item{geoBreite}{latitude of the site}

\item{geoLaenge}{longitude of the site}

\item{Hoehe_m}{height above sea level of the site}

\item{radius}{radius from which DWD weather stations should be included}

\item{minstations}{minimum number of stations to be included}

\item{startdate}{start date (string, format "1990-01-01")}

\item{maxError}{maximum deviation (%) of the values of a single weather station from the average of all stations}

\item{set_title}{title of the output file}

\item{out_file}{name of the output file}
}
\value{
none, a html file is generated
}
\description{
render_Wetter_Analysis
}
\examples{
render_Wetter_Analysis(Standort, geoBreite, geoLaenge, Hoehe_m, radius, minstations, startdate, maxError, set_title, out_file)
}
