% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderImportFunctions.R
\name{render_Wetter_Import}
\alias{render_Wetter_Import}
\title{render_Wetter_Import}
\usage{
render_Wetter_Import(
  Standort,
  geoBreite,
  geoLaenge,
  Hoehe_m,
  radius,
  minstations,
  startyear,
  endyear,
  maxError,
  set_title
)
}
\arguments{
\item{Standort}{name of the site for which weather data should be estimated (string)}

\item{geoBreite}{latitude of the site (numeric)}

\item{geoLaenge}{longitude of the site (numeric)}

\item{Hoehe_m}{height above sea level of the site (numeric)}

\item{radius}{radius from which DWD weather stations should be included (numeric)}

\item{minstations}{minimum number of stations to be included (numeric)}

\item{maxError}{maximum deviation (%) of the values of a single weather station from the average of all stations}

\item{set_title}{title of the output file}

\item{startdate}{start date (string, format "1990-01-01")}

\item{out_file}{name of the output file}
}
\value{
none, a html file is generated
}
\description{
render_Wetter_Import
}
\examples{
render_Wetter_Import(Standort, geoBreite, geoLaenge, Hoehe_m, radius, minstations, startdate, maxError, set_title, out_file)
}
