% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{ReplaceProblemChars}
\alias{ReplaceProblemChars}
\title{Title: ReplaceProblemChars}
\usage{
ReplaceProblemChars(InputString = NULL)
}
\arguments{
\item{InputString}{A string with characters to be replaced}
}
\value{
A string with replaced characters
}
\description{
Title: ReplaceProblemChars
}
\examples{
ReplaceProblemChars("Test / Test")
}
