% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{roughness_f}
\alias{roughness_f}
\title{Title roughness_f}
\usage{
roughness_f(crop_height)
}
\arguments{
\item{crop_height}{}
}
\value{
roughness_f
}
\description{
Title roughness_f
}
\examples{
roughness_f(0.1)
}
