% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{rs_f}
\alias{rs_f}
\title{Title rs_f
function for the stomatal resistance as influenced by
air temperature, net radiation, vapour pressure deficit and soil water potential
according to Neukam et al. (2018)}
\usage{
rs_f(Ta, Rnet, VPD, psi_r, rs_min)
}
\arguments{
\item{Ta}{air temperature  [°C]}

\item{Rnet}{net radiation [W/m2]}

\item{VPD}{vapour pressure deficit [hPa]}

\item{psi_r}{weighted average soil water potential in rooted soil [MPa]}

\item{rs_min}{minimum stomatal conductance [s/m]}
}
\value{
actual stomata resistance rs_f [s/m]
}
\description{
Title rs_f
function for the stomatal resistance as influenced by
air temperature, net radiation, vapour pressure deficit and soil water potential
according to Neukam et al. (2018)
}
\examples{
rs_f(20, 100, 10, 0.1, 10)
}
