% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{rs_Rnet_f}
\alias{rs_Rnet_f}
\title{Title rs_Rnet_f
function for the contribution of net radiation to stomatal resistance
according to Neukam et al. (2018)}
\usage{
rs_Rnet_f(Rnet, rs_min)
}
\arguments{
\item{Rnet}{net radiation [W/m2]}

\item{rs_min}{minimum stomatal conductance [s/m]}
}
\value{
rs_Rnet_f [s/m]
}
\description{
Title rs_Rnet_f
function for the contribution of net radiation to stomatal resistance
according to Neukam et al. (2018)
}
\examples{
rs_Rnet_f(100, 10)
}
