% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{rs_Ta_f}
\alias{rs_Ta_f}
\title{Title rs_Ta_f
function for the contribution of air temperature to stomatal resistance
according to Neukam et al. (2018)}
\usage{
rs_Ta_f(Ta, rs_min)
}
\arguments{
\item{Ta}{air temperature  [°C]}

\item{rs_min}{minimum stomatal conductance [s/m]}
}
\value{
rs_Ta_f [s/m]
}
\description{
Title rs_Ta_f
function for the contribution of air temperature to stomatal resistance
according to Neukam et al. (2018)
}
\examples{
rs_Ta_f(20, 10)
}
