% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{rs_VPD_psi_r_f}
\alias{rs_VPD_psi_r_f}
\title{Title rs_VPD_psi_r_f
function for the contribution of VPD and soil water potential to stomatal resistance
according to Neukam et al. (2018)}
\usage{
rs_VPD_psi_r_f(VPD, psi_r, rs_min)
}
\arguments{
\item{VPD}{vapour pressure deficit [hPa]}

\item{psi_r}{weighted average soil water potential in rooted soil [MPa]}

\item{rs_min}{minimum stomatal conductance [s/m]}
}
\value{
rs_VPD_psi_r_f [s/m]
}
\description{
Title rs_VPD_psi_r_f
function for the contribution of VPD and soil water potential to stomatal resistance
according to Neukam et al. (2018)
}
\examples{
rs_VPD_psi_r_f(10, 0.1, 10)
}
