% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{sat_vap_press_f}
\alias{sat_vap_press_f}
\title{Title sat_vap_press_f}
\usage{
sat_vap_press_f(Temp)
}
\arguments{
\item{Temp}{}
}
\value{
sat_vap_press_f [mbar]
}
\description{
Title sat_vap_press_f
}
\examples{
sat_vap_press_f(20)
}
