% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{SelectStations}
\alias{SelectStations}
\title{SelectStations selects the nearest weather stations from a list of stations}
\usage{
SelectStations(
  lat,
  long,
  height_loc = 100,
  stationlist,
  minstations = 7,
  max_stations = 20,
  radius = 70000,
  startdate,
  max.Height.Distance_m = 100
)
}
\arguments{
\item{lat}{Latitude of location}

\item{long}{Longitude of location}

\item{height_loc}{Height of the location}

\item{stationlist}{List of stations from which the stations should be selected}

\item{minstations}{minimum number of stations to be selected}

\item{radius}{Maximum radius where the station should be selected from}

\item{startdate}{date from which weather data should be available}

\item{max.Height.Distance_m}{maximum height}
}
\value{
data frame with DWD station info for selected stations
}
\description{
SelectStations selects the nearest weather stations from a list of stations
}
