% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{set_theme}
\alias{set_theme}
\title{Title}
\usage{
set_theme(p, fontsize)
}
\arguments{
\item{p}{a plot to be customized}

\item{fontsize}{the basic font size}
}
\value{
a plot with customized design and an adusted font size
}
\description{
Title
}
\examples{
set_theme(p, fontsize=18)
}
