% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{setplotbackground}
\alias{setplotbackground}
\title{Title}
\usage{
setplotbackground(plot, BaseSize = 18)
}
\arguments{
\item{plot}{a plot to be customized}

\item{BaseSize}{the base font size}
}
\value{
a plot with customized design and an adjusted font size
}
\description{
Title
}
\examples{
 setplotbackground(p, BaseSize=18)
}
