% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{TmaxCanopy}
\alias{TmaxCanopy}
\title{Title TmaxCanopy}
\usage{
TmaxCanopy(Tair, ra, Rn)
}
\arguments{
\item{Tair}{Air temperature [°C]}

\item{ra}{aerodynamic resistance [s/m]}

\item{Rn}{Net radiation [W/m2]}
}
\value{
TmaxCanopy [°C]
}
\description{
Title TmaxCanopy
}
\examples{
TmaxCanopy(20, 100, 600)
}
