% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{transWindSpeed}
\alias{transWindSpeed}
\title{transWindSpeed
corrects the wind speed for the required height considering
the measurement height}
\usage{
transWindSpeed(inWind, inHeight, outHeight, vegi.height)
}
\arguments{
\item{inWind}{original wind speed data [m/s]}

\item{inHeight}{measurement height of original wind speed data [m]}

\item{outHeight}{reference height to be wind speed should be converted [m]}

\item{vegi.height}{vegetation height [m]}
}
\value{
corrected wind speed data [m/s]
}
\description{
transWindSpeed
corrects the wind speed for the required height considering
the measurement height
}
\examples{
transWindSpeed(5, 10, 2, 0.5)
}
