% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weatherfunctions-package.R
\docType{package}
\name{Weatherfunctions-package}
\alias{Weatherfunctions}
\alias{Weatherfunctions-package}
\title{Weatherfunctions: Functions for downloading weather data from DWD servers}
\description{
Package for downloading weather data from DWD and some other calculations with weather data like potential evapotranspiration and canopy temperature.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://agronomykiel.github.io/Weatherfunctions/}
}

}
\author{
\strong{Maintainer}: Henning Kage \email{kage@pflanzenbau.uni-kiel.de}

Authors:
\itemize{
  \item Till Rose \email{rose@pflanzenbau.uni-kiel.de}
  \item Ingo Pahlmann
}

}
\keyword{internal}
