% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{windheight}
\alias{windheight}
\title{Alternative function for correction of windspeed to reference height}
\usage{
windheight(ui, zi, zo)
}
\arguments{
\item{ui}{original windspeed [m/s]]}

\item{zi}{measurement height [m]}

\item{zo}{reference height [m]}
}
\value{
uo corrected windspeed [m/s]
}
\description{
Alternative function for correction of windspeed to reference height
}
\examples{
windheight(5, 10, 2)
}
