% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{WriteHumeWeatherFile}
\alias{WriteHumeWeatherFile}
\title{WriteHumeWeatherFile}
\usage{
WriteHumeWeatherFile(df, fn)
}
\arguments{
\item{df}{data frame with weather data}

\item{fn}{file name for output}
}
\value{
none, a file is written
}
\description{
WriteHumeWeatherFile
}
\examples{
WriteHumeWeatherFile(df, fn)
}
