% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapFuncs.R
\name{air_saturation_deficit}
\alias{air_saturation_deficit}
\title{Title}
\usage{
air_saturation_deficit(temperature, relative_humidity)
}
\arguments{
\item{temperature}{air temperature [°C]}

\item{relative_humidity}{relative humidity [%]

#' @export}
}
\description{
Title
}
