% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{InterpolateFromDWD}
\alias{InterpolateFromDWD}
\title{Title InterpolateFromDWD
Interpolation of weather data using the 3 nearest DWD stations
in between a defined difference of altitude}
\usage{
InterpolateFromDWD(
  df_DWD_core,
  stationlist,
  geoBreite,
  geoLaenge,
  max.Height.Distance_m = 100,
  Hoehe_m = 100,
  df_Rain = NULL,
  startdate = "1990-01-01"
)
}
\arguments{
\item{df_DWD_core}{data frame with DWD weather data}

\item{stationlist}{data frame with station information}

\item{geoBreite}{Latitude of location}

\item{geoLaenge}{Longitude of location}
}
\value{
interpolated weather data in HUME format including radiation data
}
\description{
Title InterpolateFromDWD
Interpolation of weather data using the 3 nearest DWD stations
in between a defined difference of altitude
}
