% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherFunctions.R
\name{rename.weather}
\alias{rename.weather}
\title{rename weather data to long names}
\usage{
\method{rename}{weather}(DWDWeather)
}
\arguments{
\item{DWDWeather}{Data frame in DWD name style.}
}
\value{
data frame with readable column names
}
\description{
rename weather data to long names
}
\examples{
rename.weather(DWDWeather)
}
